+++ src/detection/wallpaper/wallpaper.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#ifndef FF_INCLUDED_detection_wallpaper
+#define FF_INCLUDED_detection_wallpaper
+
+#include "fastfetch.h"
+
+const char* ffDetectWallpaper(const FFinstance* instance, FFstrbuf* result);
+
+#endif