+++ src/detection/wallpaper/wallpaper_nosupport.c
@@ -1,0 +1,6 @@
+#include "wallpaper.h"
+
+const char* ffDetectWallpaper(FF_MAYBE_UNUSED const FFinstance* instance, FF_MAYBE_UNUSED FFstrbuf* result)
+{
+    return "Not supported on this platform";
+}